+++ src/model/variable.ts
