+++ src/model/view.ts
@@ -1,0 +1,7 @@
+export interface View {
+    
+}
+
+export function renderView(view: View, ctx: CanvasRenderingContext2D, strokeColor: string) {
+    
+}
