+++ src/turtle/draw.ts
@@ -1,12 +1,18 @@
-import { Turtle } from "./turtle"
-import { degrees } from "./vec2"
+import { turtle, forward } from "./turtle"
+import {Model} from '../model/model'
 
-export function draw(t: Turtle) {
+export function draw(model: Model) {
+  const t = turtle(model)
+  
+  forward(t, 100)
+}
+
+  /*
   const side = t.approxLength("A", 100)
   const angle = t.anyAngle("spin")
   const o = t.position
   t.right(angle)
-  t.right(angle)
+//  t.right(angle)
   t.forward(side)
   t.right(degrees(90))
   t.penUp()
@@ -15,7 +21,7 @@ export function draw(t: Turtle) {
 
   const d = t.position
 
-  t.atMouse(10)
+  t.atMouse(0.1)
   t.label(side)
   t.right(degrees(90))
   t.forward(side)
@@ -23,4 +29,4 @@ export function draw(t: Turtle) {
   t.forward(side)
 
   t.constrain(o, d, 100)
-}
+  */
