+++ src/ui/main.ts
@@ -3,13 +3,13 @@ import { h, render as preactRender } from "preact"
 
 import "./style.css"
 
-import * as k from "../core/api"
 import { defaultOptions } from "../core/wasmopt"
 import { Canvas } from "./canvas"
 import { checkNotNull } from "../core/assert"
 import { createPanel } from "./panel"
-import { Turtle } from "./turtle"
-import { draw } from "./draw"
+import { draw } from "../turtle/draw"
+import {Model, emptyModel, optimize} from '../model/model'
+import {renderView} from '../model/view'
 
 const html = htm.bind(h)
 
@@ -21,8 +21,8 @@ preactRender(html`<${App} />`, checkNotNull(document.getElementById("app")))
 
 const canvas = document.querySelector<HTMLCanvasElement>("#canvas")!
 
-const t = new Turtle()
-draw(t)
+let model: Model = emptyModel()
+draw(model)
 
 const ctx = canvas.getContext("2d")!
 
@@ -38,30 +38,21 @@ const config = {
 const panel = createPanel(config)
 
 function render() {
-  t.optimize(config.iterations, config.optimization)
+  model = optimize(model, config.iterations, config.optimization)
 
-  panel.render(t.computeLoss(), t.variables, k.evaluator(t.params))
+  //panel.render(model)
 
   ctx.fillStyle = config.bgColor
   ctx.fillRect(0, 0, ctx.canvas.width, ctx.canvas.height)
-  ctx.lineWidth = 1
-  ctx.strokeStyle = config.fgColor
-  t.displaySegments().forEach((s) => {
-    ctx.beginPath()
-    ctx.moveTo(s.x1, s.y1)
-    ctx.lineTo(s.x2, s.y2)
-    ctx.stroke()
-  })
 
-  t.displayLabels().forEach((l) => {
-    ctx.strokeText(l.text, l.x + 5, l.y + 10)
+  model.views.forEach((v) => {
+    renderView(v, ctx, config.fgColor)
   })
 
   requestAnimationFrame(render)
 }
 
 function handlePointerMove(e: PointerEvent) {
-  t.mouseMove(e.offsetX, e.offsetY)
 }
 
 render()
